Completely move the definitions to Internal
